requirements.txt
pycryptodome
serpent
rc2
